requirements.txt
gkeepapi==0.11.16
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
pandas==1.0.1
pandas-datareader==0.8.1